common/tmap.c: Safe comparison of two pointers.
Subtracting two pointers and assigning the result to a 32 bit integer
is never safe on 64 bit architectures -- ptrdiff_t should be used in
this situation.  Also subtracting void pointers is undefined in C.

Since we require an 'int' result, the easiest fix is to open code the
comparison we require.  GCC optimizes this function to six instructions.

The original warning was:

tmap.c: In function ‘ptrcmp’:
tmap.c:67:21: error: pointer of type ‘void *’ used in subtraction [-Werror=pointer-arith]
     return (int) (v0-v1);
                     ^

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>