icotool: Note that we are ignoring errors in palette_lookup function.
Previously palette_lookup could return -1 for error.  This was
assigned to uint32_t (thus provoking a warning from GCC), but the
error condition was never checked so 0xffffffff would end up being
assigned to the output icon.

This change just makes what we are doing here more explicit, but does
not change the semantics.

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>